Renamed `syscall::format` to `syscall::formatter`
